requirements.txt
requests
beautifulsoup4
gspread
google-auth
python-telegram-bot==20.3
os==3.13.0